+++ scripts/build-js.js
@@ -35,6 +35,7 @@ const FILES_WITHOUT_LOCAL_IMPORTS = [
   'dark-images.js',
   'copy-button.js',
   'long-progress.js',
+  'disabled-labels.js',
 ];
 
 // ---------------------------------------------------------------------------